[Admin-Panel] Made base connection
