+++ admin-panel/package.json
@@ -10,6 +10,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.5.1",
     "chart.js": "^4.4.0",
     "i18next": "^23.5.1",
     "react": "^18.2.0",
