+++ ethicapp/app.js
@@ -1,6 +1,7 @@
 "use strict";
 
 let express = require("express");
+let cors = require('cors')
 let session = require("express-session");
 let path = require("path");
 let logger = require("morgan");
@@ -13,6 +14,7 @@ let assetVersions = require("express-asset-versions");
 
 let index = require("./backend/controllers/index");
 let users = require("./backend/controllers/users");
+let admin_api = require("./backend/controllers/admin-panel-api");
 let sessions = require("./backend/controllers/sessions");
 let visor = require("./backend/controllers/visor");
 let analysis = require("./backend/controllers/analysis");
@@ -25,6 +27,8 @@ require("serve-favicon");
 require("./backend/controllers/passport-setup");
 
 let app = express();
+
+app.use(cors())
 
 //express asset versions
 var assetPath = path.join(__dirname, "/frontend/assets");
@@ -62,6 +66,7 @@ app.use(json2xls.middleware);
 
 app.use("/", index);
 app.use("/", users);
+app.use("/", admin_api);
 app.use("/", middleware.verifySession, sessions);
 app.use("/", middleware.verifySession, visor);
 app.use("/", middleware.verifySession, analysis);
